WIP: Backup local edits to operator learning
